Refactor OnBoarding Pages Descriptions
