+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.DateMe"
         tools:targetApi="31">
         <activity
-            android:name="com.androiders.dateme.features.onboading.presentation.MainActivity"
+            android:name="com.androiders.dateme.features.onboading.ui.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.DateMe">
             <intent-filter>
