--- app/src/main/java/com/androiders/dateme/features/onboading/presentation/provider/OnBoardingPages.kt
@@ -1,27 +1,0 @@
-package com.androiders.dateme.features.onboading.presentation.provider
-
-import androidx.compose.ui.graphics.Color
-import com.androiders.dateme.R
-import com.androiders.dateme.features.onboading.presentation.model.OnBoardingUiModel
-
-val page1 = OnBoardingUiModel(
-    R.drawable.coupleballoon1080w,
-    "Date Me",
-    "DateMe is a sample dating app that allows users to find others that are looking for connection. .",
-    backgroundColor = Color(0xFF0189C5),
-    mainColor = Color.White
-)
-val page2 = OnBoardingUiModel(
-    R.drawable.ballonpair1080w,
-    "A new take on an old format",
-    "The goal of this project is to practice & demonstrate the use of some Modern Android Development Practices, Material Design Theming & Jetpack Compose.",
-    backgroundColor = Color(0xFFE4AF19),
-    mainColor = Color.White
-)
-val page3 = OnBoardingUiModel(
-    R.drawable.couplefall1080w,
-    "Let’s Get Started!",
-    "Are you ready to make a profile and start exploring who is out there?",
-    backgroundColor = Color(0xFF96E172),
-    mainColor = Color.White
-)
