+++ app/src/main/java/com/androiders/dateme/features/onboading/ui/model/OnBoardingUiModel.kt
@@ -1,4 +1,4 @@
-package com.androiders.dateme.features.onboading.presentation.model
+package com.androiders.dateme.features.onboading.ui.model
 
 import androidx.compose.ui.graphics.Color
 import com.androiders.dateme.core.theme.ColorBlue
