+++ app/src/main/java/com/androiders/dateme/features/onboading/ui/provider/OnBoardingPages.kt
@@ -1,0 +1,28 @@
+package com.androiders.dateme.features.onboading.ui.provider
+
+import androidx.compose.ui.graphics.Color
+import com.androiders.dateme.R
+import com.androiders.dateme.features.onboading.ui.model.OnBoardingUiModel
+
+// TODO : Refactor this to use `stringResources`
+val page1 = OnBoardingUiModel(
+    R.drawable.coupleballoon1080w,
+    "Welcome!",
+    "Enjoy the moment. Meet new and interesting people around you.",
+    backgroundColor = Color(0xFF0189C5),
+    mainColor = Color.White
+)
+val page2 = OnBoardingUiModel(
+    R.drawable.ballonpair1080w,
+    "Discover",
+    "Find the special one. Match and connect with your type.",
+    backgroundColor = Color(0xFFE4AF19),
+    mainColor = Color.White
+)
+val page3 = OnBoardingUiModel(
+    R.drawable.couplefall1080w,
+    "Get Started!",
+    "Make a profile and start exploring who's out there.",
+    backgroundColor = Color(0xFF96E172),
+    mainColor = Color.White
+)
