+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
     <string name="app_name">DateMe</string>
     <string name="title_activity_onboarding">Onboarding</string>
+    <string name="discover">Discover</string>
 </resources>
